fix bug sidebar error
