+++ frontend/plugins/global-component.js
@@ -1,6 +1,8 @@
 import Vue from 'vue'
 import TopNavbar from '../layouts/TopNavbar'
 import ContentFooter from '../layouts/ContentFooter'
+import Sidebar from '../components/SidebarPlugin/SideBar'
+import SidebarLink from '../components/SidebarPlugin/SidebarLink'
 import ChartCard from '../components/Card/ChartCard'
 import StatsCard from '../components/Card/StatsCard'
 import NavTabsCard from '../components/Card/NavTabsCard'
@@ -8,6 +10,8 @@ import NavTabsTable from '../components/Tables/NavTabsTable'
 import OrderedTable from '../components/Tables/OrderedTable'
 import SimpleTable from '../components/Tables/OrderedTable'
 
+Vue.component("side-bar", Sidebar)
+Vue.component("sidebar-link", SidebarLink)
 Vue.component('top-navbar', TopNavbar)
 Vue.component('content-footer', ContentFooter)
 Vue.component('chart-card', ChartCard)
